OID decoding patch
Hi Graham,

I think there's a small bug in Convert-ASN1 in its decoding of OIDs - the first subidentifier isn't always being split correctly. ISO 8825-1:2003 says:
---
8.19.4 The numerical value of the first subidentifier is derived from the values of the first two object identifier components in the object identifier value being encoded, using the formula:

   (X*40)  +  Y

where X is the value of the first object identifier component and Y is the value of the second object identifier component.

NOTE – This packing of the first two object identifier components recognizes that only three values are allocated from the root node, and at most 39 subsequent values from nodes reached by X = 0 and X = 1.
---

Note the last sentence. I found an OID mentioned in Peter Gutmann's dumpasn1 program:

        OID = 06 04 86 8D 6F 63

that tickles this bug. The first subid is 86 8D 6F, which unpacks into 0189ef, which is 100079 decimal. The divide by 40 approach splits that into 2501.39. The *correct* split IMO should result in 2.99999.

The following slightly hacky patch fixes the problem, but I wanted to run it by you before committing it as it is quite surprising to see a decoding bug in a primitive...